new intersection

+++ src/main/java/org/example/project/model/Node.java
@@ -3,7 +3,7 @@ package org.example.project.model;
 public interface Node {
     long getId();
 
-    void setId(int id);
+    void setId(long id);
 
     GeoCoordinate getCoordinate();
